Change: move default impl methods in `RaftStorage` to `StorageHelper`.
- `get_initial_state()`
- `get_log_id()`
- `get_membership()`
- `last_membership_in_log()`

In the trait `RaftStorage`, these methods provide several default
methods that users do not need to care about. It should no longer
be methods that user may need to implement.

To upgrade:

If you have been using these methods, replace `sto.xxx()` with
`StorageHelper::new(&mut sto).xxx()`.
